Crate pallet_revive_eth_rpc

The EthRpcServer RPC server implementation

Modules

cli

The Ethereum JSON-RPC server.

client

The client connects to the source substrate chain and is used by the rpc server to query and send transactions to the substrate chain.

example

Example utilities

subxt_client

The generated subxt client. Generated against a substrate chain configured with pallet_revive using: subxt metadata –url ws://localhost:9944 -o rpc/revive_chain.scale

Structs

DebugRpcServerImpl

EthRpcServerImpl

An EVM RPC server implementation.

FeeHistoryProvider

Manages the fee history cache.

ReceiptExtractor

Utility to extract receipts from extrinsics.

ReceiptProvider

ReceiptProvider stores transaction receipts and logs in a SQLite database.

SubxtBlockInfoProvider

Provides information about blocks.

SystemHealthRpcServerImpl

Enums

EthRpcError

The error type for the EVM RPC server.

Constants

LOG_TARGET

Traits

BlockInfo

Provides information about a block, This is an abstratction on top of SubstrateBlock that can’t be mocked in tests. Can be removed once https://github.com/paritytech/subxt/issues/1883 is fixed.

BlockInfoProvider

BlockInfoProvider cache and retrieves information about blocks.

DebugRpcClient

Client implementation for the DebugRpc RPC API.

DebugRpcServer

Server trait implementation for the DebugRpc RPC API.

EthRpcClient

Client implementation for the EthRpc RPC API.

EthRpcServer

Server trait implementation for the EthRpc RPC API.

SystemHealthRpcClient

Client implementation for the SystemHealthRpc RPC API.

SystemHealthRpcServer

Server trait implementation for the SystemHealthRpc RPC API.