Module pallet_revive_eth_rpc::subxt_client::src_chain

Modules

alliance

alliance_motion

asset_conversion

asset_conversion_migration

asset_conversion_tx_payment

asset_rate

assets

authority_discovery

authorship

babe

balances

beefy

bounties

broker

child_bounties

contracts

conviction_voting

core_fellowship

council

democracy

election_provider_multi_phase

elections

fast_unstake

glutton

grandpa

historical

identity

im_online

indices

lottery

message_queue

mixnet

mmr

mmr_leaf

multi_block_migrations

multisig

nft_fractionalization

nfts

nis

nomination_pools

offences

pallet_example_mbms

parameters

pool_assets

pov

preimage

proxy

randomness_collective_flip

ranked_collective

ranked_polls

recovery

referenda

remark

revive

root_testing

runtime_apis

runtime_types

safe_mode

salary

scheduler

session

skip_feeless_payment

society

staking

state_trie_migration

statement

sudo

system

tasks_example

technical_committee

technical_membership

timestamp

tips

transaction_payment

transaction_storage

treasury

tx_pause

uniques

utility

verify_signature

vesting

voter_list

whitelist

Structs

ConstantsApi

CustomValuesApi

StorageApi

TransactionApi

Statics

PALLETS

RUNTIME_APIS

Functions

apis

constants

custom

is_codegen_valid_for

check whether the metadata provided is aligned with this statically generated code.

storage

tx

Type Aliases

Call

The outer extrinsic enum.

DispatchError

The error type that is returned when there is a runtime issue.

Error

The outer error enum represents the DispatchError’s Module variant.

Event

The outer event enum.