pallet_revive_eth_rpc

Trait DebugRpcServer

pub trait DebugRpcServer:
    Sized
    + Send
    + Sync
    + 'static {
    // Required methods
    fn trace_block_by_number<'life0, 'async_trait>(
        &'life0 self,
        block: BlockNumberOrTag,
        tracer_config: TracerConfig,
    ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<TransactionTrace>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn trace_transaction<'life0, 'async_trait>(
        &'life0 self,
        transaction_hash: H256,
        tracer_config: TracerConfig,
    ) -> Pin<Box<dyn Future<Output = RpcResult<Trace>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn trace_call<'life0, 'async_trait>(
        &'life0 self,
        transaction: GenericTransaction,
        block: BlockNumberOrTagOrHash,
        tracer_config: TracerConfig,
    ) -> Pin<Box<dyn Future<Output = RpcResult<Trace>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

    // Provided method
    fn into_rpc(self) -> RpcModule<Self> { ... }
}

Server trait implementation for the DebugRpc RPC API.

Required Methods

fn trace_block_by_number<'life0, 'async_trait>( &'life0 self, block: BlockNumberOrTag, tracer_config: TracerConfig, ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<TransactionTrace>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns the tracing of the execution of a specific block using its number.

References

https://geth.ethereum.org/docs/interacting-with-geth/rpc/ns-debug#debugtraceblockbynumber

fn trace_transaction<'life0, 'async_trait>( &'life0 self, transaction_hash: H256, tracer_config: TracerConfig, ) -> Pin<Box<dyn Future<Output = RpcResult<Trace>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns a transaction’s traces by replaying it.

References

https://geth.ethereum.org/docs/interacting-with-geth/rpc/ns-debug#debugtracetransaction

fn trace_call<'life0, 'async_trait>( &'life0 self, transaction: GenericTransaction, block: BlockNumberOrTagOrHash, tracer_config: TracerConfig, ) -> Pin<Box<dyn Future<Output = RpcResult<Trace>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Dry run a call and returns the transaction’s traces.

References

https://geth.ethereum.org/docs/interacting-with-geth/rpc/ns-debug#debugtracecall

Provided Methods

fn into_rpc(self) -> RpcModule<Self>

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.