Trait pallet_revive_eth_rpc::EthRpcClient

pub trait EthRpcClient: ClientT {

    // Provided methods
    fn accounts<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Address>, Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn block_number<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<U256, Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn call<'life0, 'async_trait>(
        &'life0 self,
        transaction: GenericTransaction,
        block: Option<BlockNumberOrTagOrHash>,
    ) -> Pin<Box<dyn Future<Output = Result<Bytes, Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn chain_id<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<U256, Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn estimate_gas<'life0, 'async_trait>(
        &'life0 self,
        transaction: GenericTransaction,
        block: Option<BlockNumberOrTag>,
    ) -> Pin<Box<dyn Future<Output = Result<U256, Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn gas_price<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<U256, Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_balance<'life0, 'async_trait>(
        &'life0 self,
        address: Address,
        block: BlockNumberOrTagOrHash,
    ) -> Pin<Box<dyn Future<Output = Result<U256, Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_block_by_hash<'life0, 'async_trait>(
        &'life0 self,
        block_hash: H256,
        hydrated_transactions: bool,
    ) -> Pin<Box<dyn Future<Output = Result<Option<Block>, Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_block_by_number<'life0, 'async_trait>(
        &'life0 self,
        block: BlockNumberOrTag,
        hydrated_transactions: bool,
    ) -> Pin<Box<dyn Future<Output = Result<Option<Block>, Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_block_transaction_count_by_hash<'life0, 'async_trait>(
        &'life0 self,
        block_hash: Option<H256>,
    ) -> Pin<Box<dyn Future<Output = Result<Option<U256>, Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_block_transaction_count_by_number<'life0, 'async_trait>(
        &'life0 self,
        block: Option<BlockNumberOrTag>,
    ) -> Pin<Box<dyn Future<Output = Result<Option<U256>, Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_code<'life0, 'async_trait>(
        &'life0 self,
        address: Address,
        block: BlockNumberOrTagOrHash,
    ) -> Pin<Box<dyn Future<Output = Result<Bytes, Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_storage_at<'life0, 'async_trait>(
        &'life0 self,
        address: Address,
        storage_slot: U256,
        block: BlockNumberOrTagOrHash,
    ) -> Pin<Box<dyn Future<Output = Result<Bytes, Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_transaction_by_block_hash_and_index<'life0, 'async_trait>(
        &'life0 self,
        block_hash: H256,
        transaction_index: U256,
    ) -> Pin<Box<dyn Future<Output = Result<Option<TransactionInfo>, Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_transaction_by_block_number_and_index<'life0, 'async_trait>(
        &'life0 self,
        block: BlockNumberOrTag,
        transaction_index: U256,
    ) -> Pin<Box<dyn Future<Output = Result<Option<TransactionInfo>, Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_transaction_by_hash<'life0, 'async_trait>(
        &'life0 self,
        transaction_hash: H256,
    ) -> Pin<Box<dyn Future<Output = Result<Option<TransactionInfo>, Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_transaction_count<'life0, 'async_trait>(
        &'life0 self,
        address: Address,
        block: BlockNumberOrTagOrHash,
    ) -> Pin<Box<dyn Future<Output = Result<U256, Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_transaction_receipt<'life0, 'async_trait>(
        &'life0 self,
        transaction_hash: H256,
    ) -> Pin<Box<dyn Future<Output = Result<Option<ReceiptInfo>, Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn send_raw_transaction<'life0, 'async_trait>(
        &'life0 self,
        transaction: Bytes,
    ) -> Pin<Box<dyn Future<Output = Result<H256, Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn send_transaction<'life0, 'async_trait>(
        &'life0 self,
        transaction: GenericTransaction,
    ) -> Pin<Box<dyn Future<Output = Result<H256, Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn syncing<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<SyncingStatus, Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn net_version<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<String, Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }

}

Client implementation for the EthRpc RPC API.

Provided Methods

fn accounts<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Vec<Address>, Error>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Returns a list of addresses owned by client.

fn block_number<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<U256, Error>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Returns the number of most recent block.

fn call<'life0, 'async_trait>( &'life0 self, transaction: GenericTransaction, block: Option<BlockNumberOrTagOrHash>, ) -> Pin<Box<dyn Future<Output = Result<Bytes, Error>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Executes a new message call immediately without creating a transaction on the block chain.

fn chain_id<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<U256, Error>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Returns the chain ID of the current network.

fn estimate_gas<'life0, 'async_trait>( &'life0 self, transaction: GenericTransaction, block: Option<BlockNumberOrTag>, ) -> Pin<Box<dyn Future<Output = Result<U256, Error>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Generates and returns an estimate of how much gas is necessary to allow the transaction to complete.

fn gas_price<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<U256, Error>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Returns the current price per gas in wei.

fn get_balance<'life0, 'async_trait>( &'life0 self, address: Address, block: BlockNumberOrTagOrHash, ) -> Pin<Box<dyn Future<Output = Result<U256, Error>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Returns the balance of the account of given address.

fn get_block_by_hash<'life0, 'async_trait>( &'life0 self, block_hash: H256, hydrated_transactions: bool, ) -> Pin<Box<dyn Future<Output = Result<Option<Block>, Error>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Returns information about a block by hash.

fn get_block_by_number<'life0, 'async_trait>( &'life0 self, block: BlockNumberOrTag, hydrated_transactions: bool, ) -> Pin<Box<dyn Future<Output = Result<Option<Block>, Error>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Returns information about a block by number.

fn get_block_transaction_count_by_hash<'life0, 'async_trait>( &'life0 self, block_hash: Option<H256>, ) -> Pin<Box<dyn Future<Output = Result<Option<U256>, Error>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Returns the number of transactions in a block from a block matching the given block hash.

fn get_block_transaction_count_by_number<'life0, 'async_trait>( &'life0 self, block: Option<BlockNumberOrTag>, ) -> Pin<Box<dyn Future<Output = Result<Option<U256>, Error>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Returns the number of transactions in a block matching the given block number.

fn get_code<'life0, 'async_trait>( &'life0 self, address: Address, block: BlockNumberOrTagOrHash, ) -> Pin<Box<dyn Future<Output = Result<Bytes, Error>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Returns code at a given address.

fn get_storage_at<'life0, 'async_trait>( &'life0 self, address: Address, storage_slot: U256, block: BlockNumberOrTagOrHash, ) -> Pin<Box<dyn Future<Output = Result<Bytes, Error>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Returns the value from a storage position at a given address.

fn get_transaction_by_block_hash_and_index<'life0, 'async_trait>( &'life0 self, block_hash: H256, transaction_index: U256, ) -> Pin<Box<dyn Future<Output = Result<Option<TransactionInfo>, Error>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Returns information about a transaction by block hash and transaction index position.

fn get_transaction_by_block_number_and_index<'life0, 'async_trait>( &'life0 self, block: BlockNumberOrTag, transaction_index: U256, ) -> Pin<Box<dyn Future<Output = Result<Option<TransactionInfo>, Error>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Returns information about a transaction by block number and transaction index position.

fn get_transaction_by_hash<'life0, 'async_trait>( &'life0 self, transaction_hash: H256, ) -> Pin<Box<dyn Future<Output = Result<Option<TransactionInfo>, Error>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Returns the information about a transaction requested by transaction hash.

fn get_transaction_count<'life0, 'async_trait>( &'life0 self, address: Address, block: BlockNumberOrTagOrHash, ) -> Pin<Box<dyn Future<Output = Result<U256, Error>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Returns the number of transactions sent from an address.

fn get_transaction_receipt<'life0, 'async_trait>( &'life0 self, transaction_hash: H256, ) -> Pin<Box<dyn Future<Output = Result<Option<ReceiptInfo>, Error>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Returns the receipt of a transaction by transaction hash.

fn send_raw_transaction<'life0, 'async_trait>( &'life0 self, transaction: Bytes, ) -> Pin<Box<dyn Future<Output = Result<H256, Error>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Submits a raw transaction. For EIP-4844 transactions, the raw form must be the network form. This means it includes the blobs, KZG commitments, and KZG proofs.

fn send_transaction<'life0, 'async_trait>( &'life0 self, transaction: GenericTransaction, ) -> Pin<Box<dyn Future<Output = Result<H256, Error>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Signs and submits a transaction.

fn syncing<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<SyncingStatus, Error>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Returns an object with data about the sync status or false.

fn net_version<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<String, Error>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

The string value of current network id

Object Safety

This trait is not object safe.