Trait pallet_revive_eth_rpc::ReceiptProvider

pub trait ReceiptProvider: Send + Sync {
    // Required methods
    fn insert<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        block_hash: &'life1 H256,
        receipts: &'life2 [(TransactionSigned, ReceiptInfo)],
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn remove<'life0, 'life1, 'async_trait>(
        &'life0 self,
        block_hash: &'life1 H256,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn receipt_by_block_hash_and_index<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        block_hash: &'life1 H256,
        transaction_index: &'life2 U256,
    ) -> Pin<Box<dyn Future<Output = Option<ReceiptInfo>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn receipts_count_per_block<'life0, 'life1, 'async_trait>(
        &'life0 self,
        block_hash: &'life1 H256,
    ) -> Pin<Box<dyn Future<Output = Option<usize>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn receipt_by_hash<'life0, 'life1, 'async_trait>(
        &'life0 self,
        transaction_hash: &'life1 H256,
    ) -> Pin<Box<dyn Future<Output = Option<ReceiptInfo>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn signed_tx_by_hash<'life0, 'life1, 'async_trait>(
        &'life0 self,
        transaction_hash: &'life1 H256,
    ) -> Pin<Box<dyn Future<Output = Option<TransactionSigned>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

Provide means to store and retrieve receipts.

Required Methods

fn insert<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, block_hash: &'life1 H256, receipts: &'life2 [(TransactionSigned, ReceiptInfo)], ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Insert receipts into the provider.

fn remove<'life0, 'life1, 'async_trait>( &'life0 self, block_hash: &'life1 H256, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Remove receipts with the given block hash.

fn receipt_by_block_hash_and_index<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, block_hash: &'life1 H256, transaction_index: &'life2 U256, ) -> Pin<Box<dyn Future<Output = Option<ReceiptInfo>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Get the receipt for the given block hash and transaction index.

fn receipts_count_per_block<'life0, 'life1, 'async_trait>( &'life0 self, block_hash: &'life1 H256, ) -> Pin<Box<dyn Future<Output = Option<usize>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Get the number of receipts per block.

fn receipt_by_hash<'life0, 'life1, 'async_trait>( &'life0 self, transaction_hash: &'life1 H256, ) -> Pin<Box<dyn Future<Output = Option<ReceiptInfo>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Get the receipt for the given transaction hash.

fn signed_tx_by_hash<'life0, 'life1, 'async_trait>( &'life0 self, transaction_hash: &'life1 H256, ) -> Pin<Box<dyn Future<Output = Option<TransactionSigned>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Get the signed transaction for the given transaction hash.