pallet_revive_eth_rpc::subxt_client
Module src_chain

Modules

asset_conversion

asset_conversion_migration

asset_rewards

asset_tx_payment

assets

assets_freezer

aura

aura_ext

authorship

balances

collator_selection

cumulus_xcm

foreign_assets

foreign_assets_freezer

message_queue

multi_block_migrations

multisig

nft_fractionalization

nfts

parachain_info

parachain_system

polkadot_xcm

pool_assets

pool_assets_freezer

proxy

revive

runtime_apis

runtime_types

session

state_trie_migration

system

timestamp

to_rococo_xcm_router

transaction_payment

uniques

utility

weight_reclaim

xcmp_queue

Structs

ConstantsApi

CustomValuesApi

StorageApi

TransactionApi

Statics

PALLETS

RUNTIME_APIS

Functions

apis

constants

custom

is_codegen_valid_for

check whether the metadata provided is aligned with this statically generated code.

storage

tx

Type Aliases

Call

The outer extrinsic enum.

DispatchError

The error type that is returned when there is a runtime issue.

Error

The outer error enum represents the DispatchError’s Module variant.

Event

The outer event enum.